requirements.txt
grpcio==1.65.4
grpcio-testing==1.65.4
grpcio-tools==1.65.4
protobuf==5.27.3
-e git+https://github.com/BrunoVAranha/projeto1-SD.git@559b1338e095d05f894f7c82219f04c659b6a871#egg=server
setuptools==72.1.0
